% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{covariate_removal}
\alias{covariate_removal}
\title{Covariate Correction for Gene Expression Matrix}
\usage{
covariate_removal(expression_mat, covariate_df)
}
\arguments{
\item{expression_mat}{A sample by gene numeric matrix that stores gene expression values;}

\item{covariate_df}{A sample by covariate numeric matrix that stores sample covariates;}
}
\value{
A sample by gene numeric matrix that stores corrected gene expression values.
}
\description{
Remove the effects of sample covariates from gene expression via linear regression.
}
