% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dotplot_beta_PIP}
\alias{dotplot_beta_PIP}
\title{Dotplot of the Effects of Perturbations on Factors}
\usage{
dotplot_beta_PIP(
  fit,
  target_names,
  reorder_targets = target_names,
  reorder_factors = NULL,
  exclude_offset = TRUE
)
}
\arguments{
\item{fit}{An object of class \code{"gsfa_fit"}}
}
\value{
A \code{ggplot} object.
}
\description{
Dotplot of the Effects of Perturbations on Factors
}
\details{
The size of a dot represents the PIP of association;
the color represents the effect size.
}
