% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GSFA.R
\name{fit_gsfa_multivar}
\alias{fit_gsfa_multivar}
\title{Bayesian Guided Sparse Factor Analysis on Perturbed Gene Expression Matrix}
\usage{
fit_gsfa_multivar(
  Y,
  G,
  K,
  fit0,
  init.method = c("svd", "random", "given"),
  prior_w_s = 50,
  prior_w_r = 0.2,
  prior_beta_s = 10,
  prior_beta_r = 0.2,
  niter = 500,
  average_niter = 200,
  lfsr_niter = average_niter,
  verbose = TRUE,
  return_samples = FALSE
)
}
\arguments{
\item{Y}{A numeric matrix that stores sample by gene normalized gene expression values;
\code{is.matrix(Y)} should be \code{TRUE};}

\item{G}{Either a numeric vector or a numeric matrix that stores sample-level guide information;
length or nrow of \code{G} should be the same as \code{nrow(Y)};}

\item{K}{An integer specifying the number of factors in the model; only one of \code{K}
and \code{fit0} is needed;}

\item{fit0}{A list of class 'gsfa_fit' that is obtained from a previous \code{fit_gsfa_multivar}
run, so that more iterations of Gibbs sampling can be carried out on top of it;
only one of \code{K} and \code{fit0} is needed;}

\item{init.method}{The method to initialize the factors, can be one of "svd", "random", "given";}
}
\value{
A list of class 'gsfa_fit' which stores the Gibbs sampling updates
and posterior mean estimates, and the prior parameters used during the inference.
}
\description{
Performs GSFA on given gene expression matrix and matching perturbation
(guide) information using Gibbs sampling.
}
\details{
Uses functions implemented in Rcpp from \code{GSFA_gibbs_inference.cpp}.
}
\examples{
\dontrun{
fit0 <- fit_gsfa_multivar(Y, G, 10, init.method = "svd", niter = 500, average_niter = 200)
fit1 <- fit_gsfa_multivar(Y, G, fit0 = fit0, niter = 500, average_niter = 200)
}
}
