% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{pearson_residual_transform}
\alias{pearson_residual_transform}
\title{Pearson Residual Transformation on Gene Count Matrix}
\usage{
pearson_residual_transform(count_mat)
}
\arguments{
\item{count_mat}{A sample by gene numeric matrix that stores gene count values;}
}
\value{
A sample by gene numeric matrix that stores transformed Pearson residual values.
}
\description{
Performs Pearson residual transformation (proposed in Townes et al., 2019)
and converts given raw gene count matrix into a continuous matrix with quantities
analogous to z-scores and approximately follow a normal distribution.
}
