% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{poisson_count_sim_var_scale}
\alias{poisson_count_sim_var_scale}
\title{Add a Simulated Gene Count Matrix to a Dataset with Continuous Gene
Expression}
\usage{
poisson_count_sim_var_scale(
  sim_data,
  p_scale_mean = 5e+05,
  p_scale_sd = 1e+05,
  p_offset = 1/5e+05
)
}
\arguments{
\item{sim_data}{A dataset of list type generated from \code{normal_data_sim()}}

\item{p_scale_mean}{The mean of the normal distribution where the library
size is drawn from}

\item{p_scale_sd}{The variance of the normal distribution where the library
size is drawn from}

\item{p_offset}{An offset number added to the input expression rates}
}
\value{
A list that has a sample by gene count matrix added to the original
dataset input.
}
\description{
Generate gene expression count data from a Poisson model based on
the continuous gene expression rates generated from a normal model,
incorporating varying scaling factors (library sizes) across cells.
}
\details{
Count \eqn{c_{ij}} is generated from a Poisson model based on
rate \eqn{y_{ij}} and library size (or scaling factor) \eqn{L_i} generated
from a normal model:
\deqn{L_i \sim N(\text{p_scale_mean}, \text{p_scale_sd}^2)}
\deqn{L_i = max(L_i, 0)}
\deqn{c_{ij} \sim Pois(L_i exp(y_{ij} + \text{p_offset}))}
}
\examples{
set.seed(12345)
beta_true <- rbind(c(1, 0, 0, 0, 0), c(0, 0.8, 0, 0, 0))
sim_data <- normal_data_sim(N = 4000, P = 6000, beta_true = beta_true)
sim_data <- poisson_count_sim_var_scale(sim_data)

}
