% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{select_top_devres_genes}
\alias{select_top_devres_genes}
\title{Feature Selection Based on the Deviance Statistics of Genes}
\usage{
select_top_devres_genes(dev_resid_mat, num_top_genes = 6000)
}
\arguments{
\item{dev_resid_mat}{A sample by gene numeric matrix that stores deviance residuals;}

\item{num_top_genes}{Number of top genes to keep;}
}
\value{
A numeric vector that stores the indices of top \code{num_top_genes}
genes ranked by deviance statistics.
}
\description{
Select the top genes ranked by deviance statistics.
}
