% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dotplot_beta_PIP}
\alias{dotplot_beta_PIP}
\title{Dotplot of the Effects of Perturbations on Factors}
\usage{
dotplot_beta_PIP(
  fit,
  target_names = NULL,
  reorder_targets = target_names,
  reorder_factors = NULL,
  exclude_offset = TRUE
)
}
\arguments{
\item{fit}{An object of class \code{"gsfa_fit"}}

\item{target_names}{Names of the perturbations, the default uses the
row names in \code{fit$posterior_means$beta_pm}}

\item{reorder_targets}{A subset or rearrangement of \code{target_names}
if one wishes to visualize a subset of perturbations or rearrange their order}

\item{reorder_factors}{A subset or rearrangement of factor indices if
one wishes to visualize a subset of factors or rearrange their order}
}
\value{
A \code{ggplot} object.
}
\description{
Dotplot of the Effects of Perturbations on Factors
}
\details{
The size of a dot represents the PIP of association;
the color represents the effect size.
}
