% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{dotplot_total_effect}
\alias{dotplot_total_effect}
\title{Dotplot of the Total Effects of Perturbations on Genes}
\usage{
dotplot_total_effect(
  fit,
  gene_indices,
  gene_names = gene_indices,
  target_names = NULL,
  reorder_targets = NULL,
  plot_max_score = NULL
)
}
\arguments{
\item{fit}{An object of class \code{"gsfa_fit"}}
}
\value{
A \code{ggplot} object.
}
\description{
Dotplot of the Total Effects of Perturbations on Genes
}
\details{
Sizes of the dots represent LFSR bins;
colors of the dots represent the summarized effect sizes.
}
