% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{normal_data_sim}
\alias{normal_data_sim}
\title{Simulate a dataset with gene expression and perturbation according to a normal model}
\usage{
normal_data_sim(
  N = 400,
  P = 600,
  K = 5,
  M = 2,
  beta_true = matrix(c(1, 0, 0, 0.8, 0, 0, 0, 0, 0, 0), ncol = K),
  pi_true = rep(0.1, K),
  sigma_w2_true = rep(0.5, K),
  psi_true = 1,
  G_prob = 0.05,
  offset = FALSE
)
}
\arguments{
\item{N}{Number of samples to simulate}

\item{P}{Number of genes to simulate}

\item{K}{Number of factors to simulate}

\item{M}{Number of perturbations to simulate}

\item{beta_true}{A perturbation by factor numeric matrix that stores association}

\item{pi_true}{True sparsity of each factor}

\item{G_prob}{The Bernoulli probability based on which to generate the binary
perturbation matrix G
effect sizes;}
}
\value{
A list object with the following elements:

\item{Y}{a sample by gene matrix with continuous gene
expression values;}
\item{G}{a binary sample by perturbation matrix;}
\item{Z}{a sample by factor matrix;}
\item{F}{a binary gene by factor matrix that indicates whether a gene has
non-zero loading in the factor;}
\item{U}{a gene by factor matrix with normal effect sizes, and F*U gives W,
the loading matrix.}
}
\description{
Generate a binary perturbation matrix and a continuous gene
expression matrix according to a factor model with normal noise terms.
}
