% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{poisson_count_sim_var_scale}
\alias{poisson_count_sim_var_scale}
\title{Simulate gene expression count data}
\usage{
poisson_count_sim_var_scale(
  sim_data,
  p_scale_mean = 5e+05,
  p_scale_sd = 1e+05,
  p_offset = 1/5e+05
)
}
\arguments{
\item{sim_data}{A dataset generated from \code{normal_data_sim()}}

\item{p_scale_mean}{The mean of the normal distribution where the library
size is drawn from}

\item{p_scale_sd}{The variance of the normal distribution where the library
size is drawn from}

\item{p_offset}{An offset number added to the input expression rates}
}
\value{
A list that has a sample by gene count matrix added to the original
dataset input.
}
\description{
Generate gene expression count data from a Poisson model based on
the continuous gene expression rates generated from a normal model.
}
\details{
Count \eqn{c_{ij}} is generated from a Poisson model based on
rate \eqn{y_{ij}} and library size (or scaling factor) \eqn{L_i} generated
from a normal model.
\deqn{c_{ij} \sim Pois(L_i exp(y_{ij} + \text{p_offset}))}
}
